feat: upd file structure and styles
