+++ src/app/candles/page.tsx
@@ -1,6 +1,5 @@
 import Breadcrumbs from '@components/components/Breadcrumbs/Breadcrumbs';
-import candleData from '@components/components/CandleList/candleData';
-import CandleList from '@components/components/CandleList/CandleList';
+import CandlesSection from '@components/components/CandlesSection/CandlesSection';
 
 const Candle = () => {
   return (
@@ -13,7 +12,7 @@ const Candle = () => {
           },
         ]}
       />
-      <CandleList items={candleData} />
+      <CandlesSection />
     </>
   );
 };
