+++ src/components/CandlesSection/CandleItemCard/CandleItemCard.module.css
@@ -1,0 +1,67 @@
+.card {
+  flex-basis: 154px;
+}
+
+.img_container {
+  position: relative;
+  height: 183px;
+  margin-bottom: 12px;
+}
+
+.img_container img {
+  border-radius: 20px;
+}
+
+.price {
+  font-size: 18px;
+  font-weight: 700;
+  line-height: 1.54;
+  text-transform: uppercase;
+}
+
+.price_container {
+  display: flex;
+  align-items: center;
+  gap: 6px;
+}
+
+@media (min-width: 667px) {
+  .card {
+    flex-basis: 300px;
+  }
+
+  .img_container {
+    height: 314px;
+  }
+}
+
+@media (min-width: 768px) {
+  .card {
+    flex-basis: 224px;
+  }
+
+  .img_container {
+    margin-bottom: 16px;
+    height: 248px;
+  }
+}
+
+@media (min-width: 1024px) {
+  .card {
+    flex-basis: 312px;
+  }
+
+  .img_container {
+    height: 328px;
+  }
+}
+
+@media (min-width: 1230px) {
+  .card {
+    flex-basis: 282px;
+  }
+
+  .img_container {
+    height: 314px;
+  }
+}
