+++ src/components/CandlesSection/CandleList/CandleList.module.css
@@ -1,0 +1,27 @@
+.list {
+  display: flex;
+  flex-wrap: wrap;
+  column-gap: 19px;
+  row-gap: 20px;
+}
+
+
+@media (min-width: 768px) {
+  .list {
+    column-gap: 24px;
+    row-gap: 30px;
+  }
+}
+
+@media (min-width: 1024px) {
+  .list {
+    column-gap: 20px;
+  }
+}
+
+@media (min-width: 1230px) {
+  .list {
+    column-gap: 24px;
+    row-gap: 40px;
+  }
+}
