+++ src/components/CandlesSection/CandleList/CandleList.tsx
@@ -1,0 +1,20 @@
+import { CandleI } from '../candleData';
+import CandleItemCard from '../CandleItemCard/CandleItemCard';
+
+import styles from './CandleList.module.css';
+
+interface CandleListProps {
+  items: CandleI[];
+}
+
+const CandleList: React.FC<CandleListProps> = ({ items }) => {
+  return (
+    <ul className={styles.list}>
+      {items.map((item: CandleI) => (
+        <CandleItemCard key={item.id} {...item} />
+      ))}
+    </ul>
+  );
+};
+
+export default CandleList;
