+++ src/components/CandlesSection/CandlesSection.module.css
@@ -1,0 +1,16 @@
+.section {
+  padding-top: 40px;
+}
+
+@media (min-width: 667px) {
+  .section {
+    padding-top: 50px;
+  }
+}
+
+@media (min-width: 1230px) {
+  .section {
+    padding-top: 56px;
+    padding-bottom: 120px;
+  }
+}
